CHANGE: move default parameter types to its own folder
